+++ src/pages/OrderManagement/orderList.tsx
@@ -21,8 +21,10 @@ import { AlignType } from "rc-table/lib/interface";
 import { FlexCol, FlexRow } from "../../components/Container/Container";
 import { getOrders } from "../../datasource/OrderDatasourc";
 import {
+  OrderDeliveryDestKey,
   OrderPaymentMethodName,
   OrderStatusKey,
+  ORDER_DELIVERY_DEST_METHOD_NAME,
   ORDER_PAYMENT_METHOD_NAME,
   ORDER_PAYMENT_STATUS,
   ORDER_STATUS,
@@ -77,11 +79,16 @@ export const OrderList: React.FC = () => {
   const [dateFilter, setDateFilter] = useState<any>();
   const [dataState, setDataState] = useState({
     data: [],
-    dashboard: {
-      confirmStatusCount: 0,
-      deliverySuccessCount: 0,
-      inDeliveryCount: 0,
-      waitConfirmStatusCount: 0,
+    statusCount: {
+      COMPANY_CANCEL_ORDER: 0,
+      CONFIRM_ORDER: 0,
+      DELIVERY_SUCCESS: 0,
+      IN_DELIVERY: 0,
+      OPEN_ORDER: 0,
+      REJECT_ORDER: 0,
+      SHOPAPP_CANCEL_ORDER: 0,
+      WAIT_APPROVE_ORDER: 0,
+      WAIT_CONFIRM_ORDER: 0,
     },
     count: 0,
   });
@@ -100,7 +107,7 @@ export const OrderList: React.FC = () => {
   const fetchData = async () => {
     try {
       setLoading(true);
-      const { data, dashboard, count } = await getOrders({
+      const { data, statusCount, count } = await getOrders({
         company,
         search: keyword,
         status: statusFilter,
@@ -109,8 +116,8 @@ export const OrderList: React.FC = () => {
         startDate: dateFilter && dateFilter[0] ? dateFilter[0].format("YYYY-MM-DD") : undefined,
         endDate: dateFilter && dateFilter[1] ? dateFilter[1].format("YYYY-MM-DD") : undefined,
       });
-      console.log({ dateFilter });
-      setDataState({ data, dashboard, count });
+      console.log({ data, statusCount, count });
+      setDataState({ data, statusCount, count });
     } catch (e) {
       console.log(e);
     } finally {
@@ -150,25 +157,26 @@ export const OrderList: React.FC = () => {
       color: "#FFC804",
       icon: icons.iconWaiting,
       title: "รอยืนยันคำสั่งซื้อ",
-      value: dataState.dashboard.waitConfirmStatusCount,
+      value:
+        dataState?.statusCount?.WAIT_APPROVE_ORDER + dataState?.statusCount?.WAIT_CONFIRM_ORDER,
     },
     {
       color: "#0068F4",
       icon: icons.iconCliboard,
       title: "ยืนยันคำสั่งซื้อแล้ว",
-      value: dataState.dashboard.confirmStatusCount,
+      value: dataState?.statusCount?.CONFIRM_ORDER,
     },
     {
       color: "#FF9138",
       icon: icons.iconTruck,
       title: "กำลังจัดส่ง",
-      value: dataState.dashboard.inDeliveryCount,
+      value: dataState?.statusCount?.IN_DELIVERY,
     },
     {
       color: "#2ED477",
       icon: icons.iconCheckedTruck,
       title: "จัดส่งสำเร็จ",
-      value: dataState.dashboard.deliverySuccessCount,
+      value: dataState?.statusCount?.DELIVERY_SUCCESS,
     },
   ];
 
@@ -219,13 +227,13 @@ export const OrderList: React.FC = () => {
     },
     {
       title: "การจัดส่ง",
-      dataIndex: "zone",
-      key: "zone",
+      dataIndex: "deliveryDest",
+      key: "deliveryDest",
       width: "10%",
-      render: (value: any, row: any, index: number) => {
+      render: (value: OrderDeliveryDestKey, row: any, index: number) => {
         return (
           <FlexCol>
-            <Text level={5}>{"-"}</Text>
+            <Text level={5}>{value ? ORDER_DELIVERY_DEST_METHOD_NAME[value] : "-"}</Text>
           </FlexCol>
         );
       },